progress 2 seeders
